pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "claw-log"
version = "0.1.3"
description = "Automated Career Log using generative AI and Git diffs"
readme = "README.md"
authors = [{name = "Kim Woohyuck", email = "kimwoohyuck@example.com"}]
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
]
requires-python = ">=3.7"
dependencies = [
    "google-genai>=0.3.0",
    "openai",
    "python-dotenv",
]

[project.scripts]
claw-log = "claw_log.main:main"

[tool.setuptools]
packages = ["claw_log"]
package-dir = {"claw_log" = "."}
